MAX32630FTHR: Fix NVIC_SetVector type in HAL
Change the irq_handler parameter of NVIC_SetVector from a function
pointer to a uin32_t so it matches the cmsis specification. This
prevents the following problem from occurring in RTX5 when building
for IAR:
The error: Error[Pe167]: argument of type "void (*)(void)" is incompatible
with parameter of type "uint32_t".

+++ targets/TARGET_Maxim/TARGET_MAX32630/serial_api.c
@@ -202,19 +202,19 @@ void serial_irq_set(serial_t *obj, SerialIrq irq, uint32_t enable)
 {
     switch (obj->index) {
         case 0:
-            NVIC_SetVector(UART0_IRQn, uart0_handler);
+            NVIC_SetVector(UART0_IRQn, (uint32_t)uart0_handler);
             NVIC_EnableIRQ(UART0_IRQn);
             break;
         case 1:
-            NVIC_SetVector(UART1_IRQn, uart1_handler);
+            NVIC_SetVector(UART1_IRQn, (uint32_t)uart1_handler);
             NVIC_EnableIRQ(UART1_IRQn);
             break;
         case 2:
-            NVIC_SetVector(UART2_IRQn, uart2_handler);
+            NVIC_SetVector(UART2_IRQn, (uint32_t)uart2_handler);
             NVIC_EnableIRQ(UART2_IRQn);
             break;
         case 3:
-            NVIC_SetVector(UART3_IRQn, uart3_handler);
+            NVIC_SetVector(UART3_IRQn, (uint32_t)uart3_handler);
             NVIC_EnableIRQ(UART3_IRQn);
             break;
         default:
